refactor(lottoManager): 당첨 번호에 관련된 로직을 WinningNumbers 클래스로 분리하여 처리하도록 수정
